requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-qr-code==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
nodeenv==1.6.0
oauthlib==3.2.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
segno==1.4.1
sqlparse==0.4.2
static3==0.7.0
urllib3==1.26.9
whitenoise==6.0.0
